+++ src/easy-popup_module.js
@@ -246,7 +246,7 @@ export class EasyPopup {
                     position: 'fixed',
                     display: 'none',
                     width: '100%',
-                    height: '100vh !important',
+                    height: '100vh',
                     overflowY: 'auto',
                     overflowX: 'hidden',
                     background: 'RGBA(0,0,0,0.7)',
@@ -257,6 +257,9 @@ export class EasyPopup {
                     bottom: 0,
                     textAlign: 'center',
                     verticalAlign: 'middle'
+                },
+                'body > .easy-popup-tint': {
+                    height: '100vh !important'
                 },
                 '.easy-popup_ios .easy-popup-tint': {
                     position: 'relative',
@@ -789,17 +792,24 @@ export class EasyPopup {
         // если включено закрытие попапа по клику на фон
         if (this.allowBackgroundCallback) {
             this.tintCallback = (e)=>{
-                if (!this.localFunctions.hasClass(e.target, 'easy-popup-tint'))
-                    return;
+                tint.removeEventListener('click', Popup.tintCallback, {once: true});
 
-                if (!e.target.hasAttribute('data-id') || e.target.getAttribute('data-id') != Popup.id)
+                if (!this.localFunctions.hasClass(e.target, 'easy-popup-tint')) {
+                    tint.addEventListener('click', this.tintCallback, {once: true});
                     return;
+                }
 
-                if (Popup.isLoading)
+                if (!e.target.hasAttribute('data-id') || e.target.getAttribute('data-id') != Popup.id) {
+                    tint.addEventListener('click', this.tintCallback, {once: true});
                     return;
+                }
+
+                if (Popup.isLoading || Popup.actionState) {
+                    tint.addEventListener('click', this.tintCallback, {once: true});
+                    return;
+                }
 
                 (showParams.backgroundClickCallback ? showParams.backgroundClickCallback : Popup.closerCallbackMain)();
-                tint.removeEventListener('click', Popup.tintCallback, {once: true});
             };
             tint.addEventListener('click', this.tintCallback, {once: true});
         }